chore: moved pages to Organizations folder
